# First Pass of performance improvements to Marching Tetrahedra