--- a/src/lut/mt.jl
+++ b/src/lut/mt.jl
@@ -20,98 +20,79 @@ Voxel corner and edge indexing conventions
   /
  X
 """
-struct VoxelIndices{T <: Integer}
-    voxCrnrPos::NTuple{8,Point{3,T}}
-    voxEdgeCrnrs::NTuple{19, NTuple{2,T}}
-    voxEdgeDir::NTuple{19,T}
-    voxEdgeIx::NTuple{8,NTuple{8,T}}
-    subTets::NTuple{6,NTuple{4,T}}
-    tetEdgeCrnrs::NTuple{6,NTuple{2,T}}
-    tetTri::NTuple{14,NTuple{6,T}}
 
-    function VoxelIndices{T}() where {T <: Integer}
-        voxCrnrPos = (Point(0, 0, 0),
-                      Point(0, 1, 0),
-                      Point(1, 1, 0),
-                      Point(1, 0, 0),
-                      Point(0, 0, 1),
-                      Point(0, 1, 1),
-                      Point(1, 1, 1),
-                      Point(1, 0, 1))
-        # the voxel IDs at either end of the tetrahedra edges, by edge ID
-        voxEdgeCrnrs = ((1, 2),
-                        (2, 3),
-                        (4, 3),
-                        (1, 4),
-                        (5, 6),
-                        (6, 7),
-                        (8, 7),
-                        (5, 8),
-                        (1, 5),
-                        (2, 6),
-                        (3, 7),
-                        (4, 8),
-                        (1, 3),
-                        (1, 8),
-                        (1, 6),
-                        (5, 7),
-                        (2, 7),
-                        (4, 7),
-                        (1, 7))
+const voxCrnrPos = (Point(0, 0, 0),
+                Point(0, 1, 0),
+                Point(1, 1, 0),
+                Point(1, 0, 0),
+                Point(0, 0, 1),
+                Point(0, 1, 1),
+                Point(1, 1, 1),
+                Point(1, 0, 1))
+# the voxel IDs at either end of the tetrahedra edges, by edge ID
+const voxEdgeCrnrs = ((1, 2),
+                (2, 3),
+                (4, 3),
+                (1, 4),
+                (5, 6),
+                (6, 7),
+                (8, 7),
+                (5, 8),
+                (1, 5),
+                (2, 6),
+                (3, 7),
+                (4, 8),
+                (1, 3),
+                (1, 8),
+                (1, 6),
+                (5, 7),
+                (2, 7),
+                (4, 7),
+                (1, 7))
 
-        # direction codes:
-        # 0 => +x, 1 => +y, 2 => +z,
-        # 3 => +xy, 4 => +xz, 5 => +yz, 6 => +xyz
-        voxEdgeDir = (1,0,1,0,1,0,1,0,2,2,2,2,3,4,5,3,4,5,6)
+# direction codes:
+# 0 => +x, 1 => +y, 2 => +z,
+# 3 => +xy, 4 => +xz, 5 => +yz, 6 => +xyz
+const voxEdgeDir = (1,0,1,0,1,0,1,0,2,2,2,2,3,4,5,3,4,5,6)
 
-        # For a pair of corner IDs, the edge ID joining them
-        # 0 denotes a pair with no edge
-        voxEdgeIx = ((0,1,13,4,9,15,19,14),
-                     (1,0,2,0,0,10,17,0),
-                     (13,2,0,3,0,0,11,0),
-                     (4,0,3,0,0,0,18,12),
-                     (9,0,0,0,0,5,16,8),
-                     (15,10,0,0,5,0,6,0),
-                     (19,17,11,18,16,6,0,7),
-                     (14,0,0,12,8,0,7,0))
+# For a pair of corner IDs, the edge ID joining them
+# 0 denotes a pair with no edge
+const voxEdgeIx = ((0,1,13,4,9,15,19,14),
+                (1,0,2,0,0,10,17,0),
+                (13,2,0,3,0,0,11,0),
+                (4,0,3,0,0,0,18,12),
+                (9,0,0,0,0,5,16,8),
+                (15,10,0,0,5,0,6,0),
+                (19,17,11,18,16,6,0,7),
+                (14,0,0,12,8,0,7,0))
 
-        # voxel corners that comprise each of the six tetrahedra
-        subTets = ((1,3,2,7),
-                      (1,8,4,7),
-                      (1,4,3,7),
-                      (1,2,6,7),
-                      (1,5,8,7),
-                      (1,6,5,7))
-        # tetrahedron corners for each edge (indices 1-4)
-        tetEdgeCrnrs = ((1,2),
-                        (2,3),
-                        (1,3),
-                        (1,4),
-                        (2,4),
-                        (3,4))
+# voxel corners that comprise each of the six tetrahedra
+const subTets = ((1,3,2,7),
+                (1,8,4,7),
+                (1,4,3,7),
+                (1,2,6,7),
+                (1,5,8,7),
+                (1,6,5,7))
+# tetrahedron corners for each edge (indices 1-4)
+const tetEdgeCrnrs = ((1,2),
+                (2,3),
+                (1,3),
+                (1,4),
+                (2,4),
+                (3,4))
 
-        # triangle cases for a given tetrahedron edge code
-        tetTri = ((1,3,4,0,0,0),
-                    (1,5,2,0,0,0),
-                    (3,5,2,3,4,5),
-                    (2,6,3,0,0,0),
-                    (1,6,4,1,2,6),
-                    (1,5,6,1,6,3),
-                    (4,5,6,0,0,0),
-                    (4,6,5,0,0,0),
-                    (1,6,5,1,3,6),
-                    (1,4,6,1,6,2),
-                    (2,3,6,0,0,0),
-                    (3,2,5,3,5,4),
-                    (1,2,5,0,0,0),
-                    (1,4,3,0,0,0))
-
-        new{T}(voxCrnrPos,
-            voxEdgeCrnrs,
-            voxEdgeDir,
-            voxEdgeIx,
-            subTets,
-            tetEdgeCrnrs,
-            tetTri)
-    end
-end
+# triangle cases for a given tetrahedron edge code
+const tetTri = ((1,3,4,0,0,0),
+            (1,5,2,0,0,0),
+            (3,5,2,3,4,5),
+            (2,6,3,0,0,0),
+            (1,6,4,1,2,6),
+            (1,5,6,1,6,3),
+            (4,5,6,0,0,0),
+            (4,6,5,0,0,0),
+            (1,6,5,1,3,6),
+            (1,4,6,1,6,2),
+            (2,3,6,0,0,0),
+            (3,2,5,3,5,4),
+            (1,2,5,0,0,0),
+            (1,4,3,0,0,0))