Extracting my commits from my other branch to make merging and rebase easier.

before:
  "MarchingTetrahedra{Float64}" => BenchmarkTools.Trial: 
	  memory estimate:  17.53 MiB
	  allocs estimate:  445006
	  --------------
	  minimum time:     45.394 ms (0.00% GC)
	  median time:      47.722 ms (2.93% GC)
	  mean time:        58.909 ms (3.89% GC)
	  maximum time:     981.840 ms (9.03% GC)
	  --------------
	  samples:          85
	  evals/sample:     1

after:
  "MarchingTetrahedra{Float64}" => BenchmarkTools.Trial: 
	  memory estimate:  10.27 MiB
	  allocs estimate:  120396
	  --------------
	  minimum time:     11.393 ms (0.00% GC)
	  median time:      12.211 ms (0.00% GC)
	  mean time:        13.944 ms (3.29% GC)
	  maximum time:     531.305 ms (11.73% GC)
	  --------------
	  samples:          359
	  evals/sample:     1

I have identified other performance hotspots here that bring this under the 10ms mark: https://github.com/JuliaGeometry/Meshing.jl/tree/sjk/mt-fst2
But for the test cases to pass I need parts of the workingprecision branch.